Prevent concurrent uploads of the same document by the same user
Otherwise this means that a user can try to upload a same file in
different contexts (e.g. two browsers, different browser tabs...)
and weird errors can happen.

The conflict management is already partially handled by the fact that
a client has to provide an upload offset equal to the current offset
of the resource.

The tus protocol does not have explicitly something to deal with the lock
of a resource, however official implementations handles it with a HTTP 423
code [0][1], therefore the same solution has been chosen.

This is part of story #10721: upload documents with drag'n drop

[0] https://github.com/tus/tusd/blob/0.11.0/unrouted_handler.go#L59
[1] https://github.com/tus/tus-js-client/blob/v1.5.2/lib/upload.js#L350

+++ plugins/docman/include/Tus/TusServer.php
@@ -156,15 +156,27 @@ final class TusServer implements RequestHandlerInterface
             return $this->response_factory->createResponse(400);
         }
 
-        if ((int) $request->getHeaderLine('Upload-Offset') !== $file_information->getOffset()) {
-            return $this->response_factory->createResponse(409);
-        }
+        $copied_size = 0;
+        $locker      = $this->data_store->getLocker();
+        try {
+            if ($locker !== null && ! $locker->lock($file_information)) {
+                return $this->response_factory->createResponse(423);
+            }
 
-        $copied_size = $this->data_store->getWriter()->writeChunk(
-            $file_information,
-            $file_information->getOffset(),
-            $request->getBody()->detach()
-        );
+            if ((int) $request->getHeaderLine('Upload-Offset') !== $file_information->getOffset()) {
+                return $this->response_factory->createResponse(409);
+            }
+
+            $copied_size = $this->data_store->getWriter()->writeChunk(
+                $file_information,
+                $file_information->getOffset(),
+                $request->getBody()->detach()
+            );
+        } finally {
+            if ($locker !== null) {
+                $locker->unlock($file_information);
+            }
+        }
 
         $finisher           = $this->data_store->getFinisher();
         $is_upload_finished = ($file_information->getOffset() + $copied_size) >= $file_information->getLength();
